pyproject.toml
[tool.poetry]
name = "liqui-api"
version = "0.1.0"
description = ""
authors = ["lazy-geeek <lazy-geek@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
mysql-connector-python = "^9.0.0"
gunicorn = "^23.0.0"
uvicorn = "^0.32.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
pytest-mock = "^3.14.1"
httpx = "^0.28.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true

[tool.poetry.scripts]
test = "pytest:main"